backend/requirements.txt
fastapi
uvicorn
yt-dlp
google-generativeai
python-dotenv
requests
